pyproject.toml
[tool.poetry]
name = "jwt-authentication"
version = "0.1.0"
description = ""
authors = ["ram-0209 <ramkarthick.alag@tigeranalytics.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
PyJWT = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
uvicorn = {extras = ["standard"], version = "^0.23.2"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
